Use Gradle and updated Intellij development tooling
