+++ gradle.properties
@@ -1,0 +1,3 @@
+group = com.github.leomillon.uuidgenerator
+description = UUID Generator
+version = 2.0.0
