+++ src/main/java/com/github/leomillon/uuidgenerator/ToogleDashesAction.java
@@ -1,7 +1,7 @@
-/**
+/*
  * The MIT License (MIT)
  *
- *  Copyright (c) 2015 Léo Millon
+ *  Copyright (c) 2019 Léo Millon
  *
  *  Permission is hereby granted, free of charge, to any person obtaining a copy
  *  of this software and associated documentation files (the "Software"), to deal
@@ -21,6 +21,7 @@
  *  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  *  THE SOFTWARE.
  */
+package com.github.leomillon.uuidgenerator;
 
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
